Move concrete Generating_Icon_Provider instances to new sub-namespace
