+++ includes/avatar-privacy/avatar-handlers/default-icons/generated-icons/class-wavatar-icon-provider.php
@@ -2,7 +2,7 @@
 /**
  * This file is part of Avatar Privacy.
  *
- * Copyright 2018 Peter Putzer.
+ * Copyright 2018-2019 Peter Putzer.
  *
  * This program is free software; you can redistribute it and/or
  * modify it under the terms of the GNU General Public License
@@ -24,7 +24,10 @@
  * @license http://www.gnu.org/licenses/gpl-2.0.html
  */
 
-namespace Avatar_Privacy\Avatar_Handlers\Default_Icons;
+namespace Avatar_Privacy\Avatar_Handlers\Default_Icons\Generated_Icons;
+
+use Avatar_Privacy\Avatar_Handlers\Default_Icons\Generating_Icon_Provider;
+use Avatar_Privacy\Avatar_Handlers\Default_Icons\Generators;
 
 use Avatar_Privacy\Data_Storage\Filesystem_Cache;
 
@@ -33,6 +36,7 @@ use Avatar_Privacy\Data_Storage\Filesystem_Cache;
  *
  * @since 1.0.0
  * @since 2.0.0 Moved to Avatar_Privacy\Avatar_Handlers\Default_Icons
+ * @since 2.1.0 Moved to Avatar_Privacy\Avatar_Handlers\Default_Icons\Generated_Icons
  *
  * @author Peter Putzer <github@mundschenk.at>
  */
